优化代码逻辑 1.让run函数定时10ms执行 2.取消编码器10ms定时读取，因为左右电机在都速度时有先后关系，不能同时读取，造成延迟 3.使get_track_status直接返回循迹状态值（0~63） 4.在while循环里添加了避障触发方式

+++ Hardware/encoder.h
@@ -3,7 +3,7 @@
 
 #include "main.h"
 
-int get_right_encoder_speed(void);
-int get_left_encoder_speed(void);
+short get_right_encoder_speed(void);
+short get_left_encoder_speed(void);
 
 #endif
